+++ model/mainModel.go
@@ -2,23 +2,13 @@ package model
 
 import (
 	"database/sql"
-	"fmt"
 	"log"
 	"os"
-	"strconv"
 )
 
 type ShowTodo struct { //表示用
 	Todo string
 	Id   int
-}
-
-type DelTodo struct { //削除用
-	Delete []int `form:"delete[]"`
-}
-
-type NewTodo struct { //追加用
-	Todo string `form:"new_todo"`
 }
 
 func getEnv() (string, string) { //環境変数の取得
@@ -54,31 +44,4 @@ func Select() []ShowTodo { //表示用データの作成
 		showTodos = append(showTodos, showTodo)
 	}
 	return showTodos
-}
-
-func DeleteTodo(delTodo DelTodo) {
-	db := openDB()
-	defer db.Close()
-
-	if delTodo.Delete != nil {
-		for _, v := range delTodo.Delete {
-			_, err := db.Exec("delete from todo where todoid=" + strconv.Itoa(v) + ";")
-			if err != nil {
-				log.Fatalln(err)
-			}
-		}
-	}
-}
-
-func AddTodo(newTodo NewTodo) {
-	db := openDB()
-	defer db.Close()
-	fmt.Println(newTodo.Todo)
-	if newTodo.Todo != "" {
-		_, err := db.Exec("insert into todo(todo) values('" + newTodo.Todo + "');")
-		if err != nil {
-			log.Fatalln(err)
-
-		}
-	}
 }
