+++ view/deleteView.go
@@ -1,0 +1,15 @@
+package view
+
+import (
+	"simpleTodo/model"
+
+	"github.com/gin-gonic/gin"
+)
+
+func DeleteTodo(c *gin.Context) {
+	var delTodo model.DelTodo
+	if c.ShouldBind(&delTodo) == nil {
+		model.DeleteTodo(delTodo)
+	}
+	GetRoot(c)
+}
